Handle getting directory size to enforce disk limits
This is a pretty taxing operation even on a good system, so it runs once when a file is attempting to be written, and runs using goroutines for concurrent directory listing.

Once a size is returned it is dropped into a cache and stored for 5 minutes so that subsequent write operations are faster. This cache is shared for the server as a whole, and not just the logged in user.